requirements.txt
hikari
hikari-lightbulb
pyjokes